+++ lib/glimmer/swt/composite_proxy.rb
@@ -44,6 +44,8 @@ module Glimmer
             
     end
     
+    CanvasProxy = CompositeProxy # TODO implement fully eventually
+    
   end
   
 end